orchestrator/requirements.txt
# Orchestrator Service Dependencies
# FastAPI & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP client for calling other services
httpx==0.26.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Security
slowapi==0.1.9

# Templates
jinja2==3.1.3

# Async support
anyio==4.2.0

# Logging
structlog==24.1.0

# Prometheus metrics
prometheus-client==0.19.0

# Testing (dev)
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
